Support for Symfony 4
