+++ .travis.yml
@@ -2,15 +2,23 @@ language: php
 
 sudo: false
 
-matrix:
-  fast_finish: true
-
-before_script:
-  - composer self-update
-  - composer install --prefer-source
-
 php:
   - 5.5
   - 5.6
   - 7.0
+  - 7.1
   - hhvm
+
+matrix:
+  include:
+    - php: 5.5
+      env: COMPOSER_OPTIONS="--prefer-lowest --prefer-stable"
+    - php: 7.0
+      env: COMPOSER_REQUIRE="symfony/process:^2.0"
+    - php: 7.0
+      env: COMPOSER_REQUIRE="symfony/process:^3.0"
+
+before_script:
+  - composer self-update
+  - if [ -n "$COMPOSER_REQUIRE" ]; then composer require --no-update $COMPOSER_REQUIRE; fi
+  - composer update $COMPOSER_OPTIONS
