+++ composer.json
@@ -25,10 +25,10 @@
         "evenement/evenement" : "^2.0|^1.0",
         "monolog/monolog"     : "^1.3",
         "psr/log"             : "^1.0",
-        "symfony/process"     : "^2.0|^3.0"
+        "symfony/process"     : "^2.3|^3.0|^4.0"
     },
     "require-dev": {
-        "phpunit/phpunit"     : "^4.0"
+        "phpunit/phpunit"     : "^4.0|^5.0"
     },
     "autoload": {
         "psr-0": {
